refactor Translator.py in ltlf2dfa.py, add to_dfa method to LTLf and PLTLf formula
